[UOT-139911] Update to make entities_dict a global accessible variable from within entities.py - seperate out utils into seperate file.
